Updated peerData.online parsing
Set peerData.online to 0 if peer is not found in peers list
Set peerData.online to 1 if peer is found in peers list and latency is not -1 (i.e. a direct route exists)
Set peerData.online to 2 if peer is found in peers list and latency is -1 (i.e. relayed connection)